Initial conversion to lazy.nvim
