--- config/nvim/after/ftplugin/markdown.lua
@@ -1,2 +1,0 @@
--- add bindings for nmap <C-p> <Plug>MarkdownPreviewToggle
--- add bindi ngs for command :Glow
