--- config/nvim/lua/etiennecollin/config/bufferline.lua
@@ -1,5 +1,0 @@
-require("bufferline").setup {
-    options = {
-        diagnostics = "nvim_lsp"
-    }
-}
