--- config/nvim/lua/etiennecollin/config/markdownpreview.lua
@@ -1,4 +1,0 @@
-vim.g.mkdp_auto_start = 0
-vim.g.mkdp_auto_close = 1
-vim.g.mkdp_echo_preview_url = 1
-vim.g.mkdp_page_title = "${name}.md"
