--- config/nvim/lua/etiennecollin/config/toggleterm.lua
@@ -1,1 +1,0 @@
-require("toggleterm").setup()
