+++ config/nvim/lua/etiennecollin/functions.lua
@@ -1,11 +1,8 @@
-function setTheme(theme)
+function set_theme(theme)
     -- Set theme
-    theme = theme or "dracula"
     vim.cmd.colorscheme(theme)
 
     -- Set transparency
     -- vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
     -- vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
 end
-
-setTheme()
