+++ config/nvim/lua/etiennecollin/plugins/dracula.lua
@@ -1,0 +1,5 @@
+return {
+    "Mofiqul/dracula.nvim",
+    priority = 1000,
+    config = function () set_theme("dracula") end,
+}
