+++ config/nvim/lua/etiennecollin/plugins/fugitive.lua
@@ -1,0 +1,1 @@
+return {"tpope/vim-fugitive"}
