+++ config/nvim/lua/etiennecollin/plugins/harpoon.lua
@@ -1,0 +1,4 @@
+return {
+    "theprimeagen/harpoon",
+    dependencies = "nvim-lua/plenary.nvim"
+}
