+++ config/nvim/lua/etiennecollin/plugins/lualine.lua
@@ -1,0 +1,11 @@
+return {
+    "nvim-lualine/lualine.nvim",
+    dependencies = "nvim-tree/nvim-web-devicons",
+    config = function()
+        require("lualine").setup {
+            options = {
+                theme = "dracula"
+            }
+        }
+    end
+}
