+++ config/nvim/lua/etiennecollin/plugins/markdown-preview.lua
@@ -1,0 +1,14 @@
+return {
+    "iamcco/markdown-preview.nvim",
+    cmd = {"MarkdownPreviewToggle", "MarkdownPreview", "MarkdownPreviewStop"},
+    ft = {"markdown"},
+    build = function()
+        vim.fn["mkdp#util#install"]()
+    end,
+    init = function()
+        vim.g.mkdp_auto_start = 0
+        vim.g.mkdp_auto_close = 1
+        vim.g.mkdp_echo_preview_url = 1
+        vim.g.mkdp_page_title = "${name}.md"
+    end
+}
