+++ config/nvim/lua/etiennecollin/plugins/telescope.lua
@@ -1,0 +1,5 @@
+return {
+    "nvim-telescope/telescope.nvim",
+    branch = "0.1.x",
+    dependencies = "nvim-lua/plenary.nvim"
+}
