+++ config/nvim/lua/etiennecollin/plugins/toggleterm.lua
@@ -1,0 +1,5 @@
+return {
+    "akinsho/toggleterm.nvim",
+    version = "*",
+    config = true
+}
