+++ config/nvim/lua/etiennecollin/plugins/trouble.lua
@@ -1,0 +1,9 @@
+return {
+    "folke/trouble.nvim",
+    dependencies = {"nvim-tree/nvim-web-devicons"},
+    opts = {
+        -- your configuration comes here
+        -- or leave it empty to use the default settings
+        -- refer to the configuration section below
+    }
+}
