+++ config/nvim/lua/etiennecollin/plugins/undotree.lua
@@ -1,0 +1,1 @@
+return {"mbbill/undotree"}
